+++ CHANGELOG.md
@@ -127,6 +127,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Updated `meta.total` to `meta.total_pages` in preprint-form-authors component
 - Modified `preprint-status-banner` component for review action rename.
 - Update to ember-osf@0.13.0
+- Update preprints to work with a new unified version of ember-osf (works with both preprints and reviews apps)
 
 ### Fixed
 - component integration tests to work in Firefox